requirements.txt
flask==2.2.3
requests==2.28.2
python-dotenv==1.0.0
spotipy==2.22.1
serpapi==0.1.0
gunicorn==20.1.0
